requirements.txt
Flask
tensorflow
keras
joblib
pandas
numpy
google-cloud-storage
